// src/components/neon-crm/types.ts
export type EmailAndType = {
  email: string;
  type: "Work" | "Home" | "Other";
};

export type PhoneNumberAndType = {
  number: string;
  type: "Work" | "Home" | "Other";
};

export type Contact = {
  first_name: string;
  last_name: string;
  title: string;
  company_id?: Identifier | null;
  email_jsonb: EmailAndType[];
  avatar?: Partial<RAFile>;
  linkedin_url?: string | null;
  first_seen: string;
  last_seen: string;
  has_newsletter: boolean;
  tags: number[];
  gender: string;
  sales_id?: Identifier;
  status: string;
  source?: string;
  background: string;
  phone_jsonb: PhoneNumberAndType[];
  nb_tasks?: number;
  company_name?: string;
  contact_status_id?: string | null;
  contact_status?: ContactStatus;
} & Pick<RaRecord, "id">;
// ...
export type ContactStatus = {
  tenant_id: string;
  label: string;
  color: string;
  icon?: string;
  position: number;
  is_default: boolean;
  created_at: string;
} & Pick<RaRecord, "id">;
// ...
export interface RAFile {
  src: string;
  title: string;
  path?: string;
  rawFile: File;
  type?: string;
}